refactor: Notify previous fragment of post creation
When a post is successfully created in `ContentsCreateFragment`, this change ensures that the previous fragment in the backstack (expected to be `PlazaFragment`) is notified.

This is achieved by setting a value in the `savedStateHandle` of the previous backstack entry. `PlazaFragment` can then observe this value to refresh its content or perform other actions upon returning.